refactor: replace Perceptron with PerceptronLite by deleting several unused methods
